The problem was not the wall. The issue was in the loop and mask.

+++ src/config/config.h
@@ -3,7 +3,7 @@
 #include <cstdint>
 
 // Defining some macro for a easy change when needed
-using varType = float;
+using varType = double;
 using CInt = const int;
 using maskType = uint8_t;
 using nodeType = uint8_t;
